Update
Functionality for unzipping and extracting information from EU4 savegames. Currently only extracts information related to province ownership changes.

Various updates, such as fewer variables floating around by moving some of the defines to dictionaries.

The script can now load province.png and analyze it to figure out which pixels on the province belong to which provinces. Some code for using this to produce a map of the world based on owners - if any - of provinces has been added.